+++ WarpBusiness.Web/Program.cs
@@ -144,6 +144,9 @@ builder.Services.AddHttpClient<CrmApiClient>(configureApiClient)
 builder.Services.AddHttpClient<CatalogApiClient>(configureApiClient)
     .AddHttpMessageHandler<AuthTokenHandler>();
 
+builder.Services.AddHttpClient<TaxonomyApiClient>(configureApiClient)
+    .AddHttpMessageHandler<AuthTokenHandler>();
+
 builder.Services.AddSingleton<CatalogTemplateRenderer>();
 
 builder.Services.AddRazorComponents()